+++ tests/identity_test.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-# TODO: remove once there are unit tests in repo
-def test_identiy():
-    assert 1 == 1